Specify versions which will build on windows Test CLI on all major platforms.
