+++ .github/workflows/build-rust.yml
@@ -26,7 +26,11 @@ on:
 jobs:
   rust:
     name: Test
-    runs-on: ubuntu-latest
+    runs-on: ${{ matrix.os }}
+    strategy:
+      fail-fast: false
+      matrix:
+        os: [ ubuntu-latest, windows-latest, macos-latest ]
     defaults:
       run:
         working-directory: ./cli
